Move all constraints to root constraints directory
Move all constraint files from their category subdirectories back to the
root constraints/ directory. This will give us flexibility to discover
more natural groupings based on real-world usage patterns before
finalizing categorical organization.

Changelog:
  * moved all *_constraint.rb files from comparison/, logical/,
    and structural/ subdirectories to constraints/
  * removed empty category subdirectories

The implementation and behavior of all constraints remains unchanged;
this is purely an organizational change to defer categorization until
we have more usage data.
