+++ domainic-type/lib/domainic/type/constraint/constraints/all_constraint.rb
