+++ domainic-type/lib/domainic/type/constraint/constraints/and_constraint.rb
