+++ domainic-type/lib/domainic/type/constraint/constraints/any_constraint.rb
