+++ domainic-type/lib/domainic/type/constraint/constraints/emptiness_constraint.rb
