+++ domainic-type/lib/domainic/type/constraint/constraints/equality_constraint.rb
