+++ domainic-type/lib/domainic/type/constraint/constraints/none_constraint.rb
