+++ domainic-type/lib/domainic/type/constraint/constraints/nor_constraint.rb
