+++ domainic-type/lib/domainic/type/constraint/constraints/not_constraint.rb
