+++ domainic-type/lib/domainic/type/constraint/constraints/or_constraint.rb
