+++ domainic-type/lib/domainic/type/constraint/constraints/ordering_constraint.rb
