+++ domainic-type/lib/domainic/type/constraint/constraints/type_constraint.rb
