+++ domainic-type/lib/domainic/type/constraint/constraints/uniqueness_constraint.rb
