+++ domainic-type/sig/domainic/type/constraint/constraints/all_constraint.rbs
