+++ domainic-type/sig/domainic/type/constraint/constraints/and_constraint.rbs
