+++ domainic-type/sig/domainic/type/constraint/constraints/any_constraint.rbs
