+++ domainic-type/sig/domainic/type/constraint/constraints/emptiness_constraint.rbs
