+++ domainic-type/sig/domainic/type/constraint/constraints/equality_constraint.rbs
