+++ domainic-type/sig/domainic/type/constraint/constraints/none_constraint.rbs
