+++ domainic-type/sig/domainic/type/constraint/constraints/nor_constraint.rbs
