+++ domainic-type/sig/domainic/type/constraint/constraints/not_constraint.rbs
