+++ domainic-type/sig/domainic/type/constraint/constraints/type_constraint.rbs
